requirements.txt
torch>=2.2.0
numpy<2
flask>=2.3.0
chromadb>=0.4.0
pymongo>=4.6.0
python-dotenv>=1.0.0
ollama>=0.1.0
sentence-transformers==2.2.2
transformers>=4.35.0,<4.40.0
requests>=2.31.0
